requirements.txt
astroid==2.0
colorama==0.3.9
Django==1.11
django-filter==2.0.0
django-shortcuts==1.5
djangorestframework==3.8.2
flake8==3.5.0
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
mpmath==1.0.0
psycopg2==2.7.1
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==2.0.0
pytz==2018.4
selenium==3.13.0
six==1.11.0
sympy==1.2
typed-ast==1.1.0
wrapt==1.10.11
